add comment reply functions

+++ forum-api/src/main/java/com/platform/entity/ReportVo.java
@@ -6,11 +6,11 @@ public class ReportVo implements Serializable {
     private static final long serialVersionUID = 1L;
 
     //主键
-    private Integer oId;
+    private Long oId;
     //举报人 id
-    private Integer reportUserId;
+    private Long reportUserId;
     //举报数据 id
-    private Integer reportDataId;
+    private Long reportDataId;
     //0：帖子，1：回帖，2：用户
     private Integer reportDataType;
     //0：（内容）垃圾广告，1：（内容）色情低俗，2：（内容）违法违规，3：（内容）涉嫌侵权，4：（内容）人身攻击，5：（用户）冒充账号，6：（用户）垃圾广告账号，7：（用户）个人信息违规，49：其他
@@ -20,27 +20,27 @@ public class ReportVo implements Serializable {
     //0：未处理，1：已处理
     private Integer reportHandled;
 
-    public Integer getoId() {
+    public Long getoId() {
         return oId;
     }
 
-    public void setoId(Integer oId) {
+    public void setoId(Long oId) {
         this.oId = oId;
     }
 
-    public Integer getReportUserId() {
+    public Long getReportUserId() {
         return reportUserId;
     }
 
-    public void setReportUserId(Integer reportUserId) {
+    public void setReportUserId(Long reportUserId) {
         this.reportUserId = reportUserId;
     }
 
-    public Integer getReportDataId() {
+    public Long getReportDataId() {
         return reportDataId;
     }
 
-    public void setReportDataId(Integer reportDataId) {
+    public void setReportDataId(Long reportDataId) {
         this.reportDataId = reportDataId;
     }
